Creating PromptConfig
