+++ aiperf/common/config/input/__init__.py
@@ -4,15 +4,11 @@
 __all__ = [
     "AudioConfig",
     "ImageConfig",
-    "InputConfig",
-    "OutputTokensConfig",
+    "PromptConfig",
     "SessionsConfig",
-    "SyntheticTokensConfig",
 ]
 
 from aiperf.common.config.input.audio_config import AudioConfig
 from aiperf.common.config.input.image_config import ImageConfig
-from aiperf.common.config.input.input_config import InputConfig
-from aiperf.common.config.input.output_tokens_config import OutputTokensConfig
+from aiperf.common.config.input.prompt_config import PromptConfig
 from aiperf.common.config.input.sessions_config import SessionsConfig
-from aiperf.common.config.input.synthetic_tokens_config import SyntheticTokensConfig
